+++ example/lib/pages/mini_games/multi_scenario/maps/map_biome_1.dart
@@ -18,7 +18,6 @@ class MapBiome1 extends StatelessWidget {
   Widget build(BuildContext context) {
     return BonfireWidget(
       joystick: Joystick(
-        keyboardConfig: KeyboardConfig(),
         directional: JoystickDirectional(),
       ),
       player: GamePlayer(
