+++ example/lib/pages/mini_games/multi_scenario/maps/map_biome_2.dart
@@ -20,7 +20,6 @@ class MapBiome2 extends StatelessWidget {
   Widget build(BuildContext context) {
     return BonfireWidget(
       joystick: Joystick(
-        keyboardConfig: KeyboardConfig(),
         directional: JoystickDirectional(),
       ),
       player: GamePlayer(
