+++ example/lib/pages/mini_games/top_down_game/top_down_game.dart
@@ -30,7 +30,6 @@ class TopDownGame extends StatelessWidget {
         },
       ),
       joystick: Joystick(
-        keyboardConfig: KeyboardConfig(),
         directional: JoystickDirectional(),
         actions: [
           JoystickAction(
