+++ lib/input/input.dart
@@ -1,5 +1,5 @@
 export 'package:bonfire/input/gestures/drag_gesture.dart';
 export 'package:bonfire/input/gestures/gesture_event.dart';
 export 'package:bonfire/input/gestures/tap_gesture.dart';
-export 'package:bonfire/input/keyboard_listener.dart';
+export 'package:bonfire/input/keyboard/keyboard_listener.dart';
 export 'package:bonfire/input/mouse_listener.dart';
