+++ lib/widgets/bonfire_widget.dart
@@ -7,14 +7,15 @@ import 'package:bonfire/camera/camera_config.dart';
 import 'package:bonfire/color_filter/game_color_filter.dart';
 import 'package:bonfire/forces/forces_2d.dart';
 import 'package:bonfire/game_interface/game_interface.dart';
-import 'package:bonfire/joystick/joystick_controller.dart';
+import 'package:bonfire/input/keyboard/keyboard_config.dart';
+import 'package:bonfire/input/player_controller.dart';
 import 'package:bonfire/map/base/map_game.dart';
 import 'package:bonfire/player/player.dart';
 import 'package:flutter/material.dart';
 
 class BonfireWidget extends StatefulWidget {
   /// The player-controlling component.
-  final JoystickController? joystick;
+  final PlayerController? joystick;
 
   /// Represents the character controlled by the user in the game. Instances of this class has actions and movements ready to be used and configured.
   final Player? player;
@@ -60,11 +61,13 @@ class BonfireWidget extends StatefulWidget {
   final GameColorFilter? colorFilter;
   final VoidCallback? onDispose;
   final List<Force2D>? globalForces;
+  final KeyboardConfig? keyboardConfig;
 
   const BonfireWidget({
     Key? key,
     required this.map,
     this.joystick,
+    this.keyboardConfig,
     this.player,
     this.interface,
     this.background,
